Integracion Spring Security
1.-)Integracion de formulario de autenticacion con spring security.
2.-) Mensajes dinamicos.
3.-) Manejo de Errores.
4.-) Manejo y control de sesion.

+++ ISoft-Profile-Api/src/main/java/co/isoft/nnita/profile/api/models/Usuarios.java
@@ -66,7 +66,7 @@ public class Usuarios implements Serializable,BusinessClass
     private Date fecha_ultima_visita;
 
     @Column(name = "HABILITADO", nullable = false)
-    private int habilitado;
+    private Long habilitado;
 
     public Long getId()
     {
@@ -158,12 +158,12 @@ public class Usuarios implements Serializable,BusinessClass
         this.fecha_ultima_visita = fecha_ultima_visita;
     }
 
-    public int getHabilitado()
+    public Long getHabilitado()
     {
         return habilitado;
     }
 
-    public void setHabilitado(int habilitado)
+    public void setHabilitado(Long habilitado)
     {
         this.habilitado = habilitado;
     }
